kola/tests/*: use c.SSH in place of m.SSH
This replaces all uses of `m.SSH` with `c.SSH`.

This allows all uses of `SSH` from a given test to produce log output
actually associated with that test rather than just written straight to
stderr.

This refactor was accomplished mostly with sed, though helper functions
had to be manually tweaked.

The reduction in linecount is due to also refactoring the
`genDockerContainer` function, now that it takes a 'c', to just
`c.Fatalf` directly rather than return an error.